WIP plot CFG.
